+++ CHANGELOG.md
@@ -12,6 +12,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3486 Add physical activity name in tooltip
 - YLP-3522 Handle accounts marked for deletion
 - YLP-3549 Update dashboard cards UI
+- YLP-3574 Meal bolus correction display
 - YLP-3636 Update banner to display HTML content
 
 ### Fixed